feat: license check
