+++ package.json
@@ -8,12 +8,14 @@
     "@hapi/hapi": "20.2.1"
   },
   "devDependencies": {
+    "@feasibleone/aegis": "^1.0.0",
     "tap": "^16.0.0",
     "ut-function.capture-hapi": "^1.1.3",
     "ut-run": "^10.69.8",
     "ut-port-script": "^7.2.0"
   },
   "peerDependencies": {
+    "@feasibleone/aegis": "^1.0.0",
     "ut-port-script": "^7.1.1",
     "ut-bus": "^7.47.0",
     "ut-run": "^10.64.6"
